Refactored to use TlsHandler
