+++ pom.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <groupId>com.github.chhsiao90</groupId>
     <artifactId>nitmproxy</artifactId>
@@ -26,9 +25,7 @@
         <assertj.version>3.18.1</assertj.version>
         <mockito.version>3.6.28</mockito.version>
         <jetty.alpnAgent.version>2.0.10</jetty.alpnAgent.version>
-        <jetty.alpnAgent.path>
-            "${settings.localRepository}"/org/mortbay/jetty/alpn/jetty-alpn-agent/${jetty.alpnAgent.version}/jetty-alpn-agent-${jetty.alpnAgent.version}.jar
-        </jetty.alpnAgent.path>
+        <jetty.alpnAgent.path>"${settings.localRepository}"/org/mortbay/jetty/alpn/jetty-alpn-agent/${jetty.alpnAgent.version}/jetty-alpn-agent-${jetty.alpnAgent.version}.jar</jetty.alpnAgent.path>
         <argLine.alpnAgent>-javaagent:${jetty.alpnAgent.path}</argLine.alpnAgent>
         <argLine.leak>-D_</argLine.leak>
         <argLine.coverage>-D_</argLine.coverage>
@@ -71,7 +68,6 @@
                     </execution>
                 </executions>
             </plugin>
-
             <!-- we can either fork it as a seperate JVM or run it within the same process -->
             <plugin>
                 <groupId>org.codehaus.mojo</groupId>
@@ -118,7 +114,6 @@
                     </execution>
                 </executions>
             </plugin>
-
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-surefire-plugin</artifactId>
@@ -140,8 +135,7 @@
         <profile>
             <id>leak</id>
             <properties>
-                <argLine.leak>-Dio.netty.leakDetectionLevel=paranoid -Dio.netty.leakDetection.maxRecords=32
-                </argLine.leak>
+                <argLine.leak>-Dio.netty.leakDetectionLevel=paranoid -Dio.netty.leakDetection.maxRecords=32</argLine.leak>
             </properties>
         </profile>
         <profile>
