+++ src/main/java/com/github/chhsiao90/nitmproxy/ConnectionContext.java
@@ -73,7 +73,7 @@ public class ConnectionContext {
             case SOCKS:
                 return new SocksProxyHandler(master, this);
             case TRANSPARENT:
-                return new TransparentProxyHandler(master, this);
+                return new TransparentProxyHandler(this);
             default:
                 throw new IllegalStateException("No proxy mode available: " + master.config().getProxyMode());
         }
